Add OffenderScraper and associated cache model
This adds a module (OffenderScraper) that uses mechanize, a HTTP
scraping library, to fetch details about an offender from the public
Oregon Offender Search (OOS) interface. Since we don't have access to
this data in a programmatic fashion, this will have to do for now.

Since it takes about a second to fetch this data, and to be polite to
the OOS servers, this commit also adds a 1 day cache that is persisted
in the database in the `offender_search_caches` table. This cache is
never evicted, I don't believe we will ever need to clear it out since
the number of offenders is not growing quickly.