Add OffenderSearchCache.purge_all! method
This method purges whatever search caches happen to be in the query
scope. Useful for manual operations whenever there is a new field added
to the cache.

This also moves the AGE_LIMIT into the model as a default scope. This is
pretty sneaky, but since the model is pretty clearly a read-through
cache, hopefully this is less unexpected.