Add an error state for when the offender is not found
This gives the user an error, kind of like a 404, but on the search
page. This is not necessarily the ultimate place for this, but I wanted
to push this to get feedback.

The error is tracked in mixpanel under 'message-seen', and has the
'level: error' attribute so it should be possible to see which users get
which errors.

Note that there are two conditions rolled up into the same error message
currently - when an invalid SID is entered (i.e. non numeric) AND when a
SID that doesn't correspond to an inmate is entered.

The actual error message:

    We couldn't find that offender. You may want to double-check the SID
    you used to search.

    We currently only have information on offenders in the prison
    system–if the offender with SID [...SID...] is not currently in a
    state prison then we will not have their information.