Remove notification_id from sessions
The smart links are a bust. It's too hard to explain what they do (even
1-on-1 with people not in trauma) and it makes it super hard to navigate
through the site because you (currently) have to clear cookies in order
to get the link out of the session.

Anyone clicking through a legacy link will be redirected based on the
value of the "o" param (the SID).