Add initial version of style guide
This is a first cut at a style guide located at /styleguide and
introduces a single component as a starting example: the "dropdown"
component.

This component accepts two options -- the "title" and the dropdown body,
and can be rendered in a view like so:

```haml
= render_component 'dropdown', title: 'Dropdown Title Here' do
  This is the body of the dropdown!
```

As for adding new components, for a component named "some_component",
follow these filename conventions:

  app/assets/components/some_component
                        ├── index.css.scss
                        └── _index.html.haml    (note the underscore)

To add an instance of it in the styleguide, add the render_component
call to app/views/styleguides/show.html.haml.