Extract typography into a placeholder mixin
This has the benefit that this file can be included by any component and
the styles in that component can extend the typographical styles. It's
important that the typography.scss mixin not output any styles -- since
it is included by multiple files, any styles that are output in this
file will be repeated once per inclusion.

This also removes the custom "subheader" typographical style in favor of
the new header-3 style which is similar enough during the transition.