Update <input> font to use body-2
This is tricky because the selectors in the stylesheet are lower
precedence than the ones materialize uses. So, the font-size in
materialize overrides the font-size coming in from the mixin (hence the
override of $input-font-size). Tweaking the line-height was necessary so
the label wasn't too far from the text input.