Add FAQ page scaffolding
There isn't too much content in here yet (only the menu items). But this
starts a pattern of defining the FAQ items in a constant in the
controller and then matching up each item to an I18n key for rendering.

The URL to the page is dynamically generated based on the menu item
text. For example "Questions before trial" will have the URL
/faqs/questions-before-trial.