Don't show "no results" when there is a connection error
Due to the complex conditionals in this controller method, when there
was a connection error, the results array would be set to an empty array
and then the "no results" error would be displayed instead of the
connection error.

I think this fixes it without breaking the other cases, but it's hard to
test when OOS is down.