Remove SCSS import of "materialize/components/color"
The variables specified in this SCSS file are required by other
components that materialize provides (like
"materialize/components/variables"), so any of our top-level stylesheets
need to include that SCSS file. The problem is that this "color" SCSS
file also outputs a lot of CSS rules (two for each color variant) which
would be duplicated in any top-level stylesheet that includes the SCSS
file, so it becomes wasteful to require this file in multiple top-level
stylesheets. If only we could just get the variables without the styling
rules...

To make this happen, let's just set the variables that we need in our
own colors mixin.

Some of these colors appear to be unused as of currently, so I've made
them an ugly red so we notice them in case they happen to be used.

This also adds a full-bleed class for the FAQ page to demonstrate how we
will need to include the variables in another place.