Use the DA Victim Services phone number for contact info
I couldn't find a good digital way to enable these communications -- no
email address or anything on their website. Maybe there is still some
way we can do this... but in the spirit of shipping something early and
polishing it later, let's get this workable with a simple phone number.