Remove default parameter assignment from toggleDropdown
Okay, I can't figure this one out and I've gotta go, so a quick fix for
now.

The problem is that there is a syntax error processing the default
parameter assignment. MDN docs say that this syntax is not universally
supported[1].

Rails uses the Uglifier gem to do javascript minification for
production, which in turn uses execjs to shell out to an underlying
javascript engine.

I can reproduce this with:

  > Uglifier.compile "const _ = function(foo = 1) { return foo; }"
  ExecJS::RuntimeError: SyntaxError: Unexpected token operator «=», expected punc «,»

However, the underlying ExecJS call works just fine:

  > context = ExecJS.compile "const _ = function(foo = 1) { return foo; }"
  > context.call("_")
  => 1

It must be one of the uglifier settings passed to ExecJS#call that
triggers this behavior.

The ideal solution here is to not use uglifier, rather to use a babel
workflow to compile ES6 features like "const" into universally supported
syntax.

[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Functions/Default_parameters#Browser_compatibility