Update FAQ menu links to be primary color instead
The default link styling (colored the same as text, but underlined)
looks not good in the menu. Using the primary color seems to work better
here.

Note that previously, these links had been colored based off the default
materialize $primary-color value, which was a light blue, that had
nothing to do with our color pallet. So this coloration is a bit darker,
and maybe we'll want to re-introduce a lighter variant to make these
links stand out again, but we should do that intentionally and not
because we accidentally inherited a default from materialize.css.