Add flipper feature flag library and `log_in` flag
This adds a lightweight library that has some pretty good support for
advanced feature flag flipping techniques that we will probably never
need.

The key part is that it abstracts the conditional enabling of a feature
from environment variables by one level without tying it to the Rails
environment.

For example, we avoid this antipattern:

```ruby
- if !Rails.env.production?
  Hidden Stuff Here
```

This bites when in the test suite (`Rails.env.test?`) and suddenly you
have to stub Rails.env in order to test your feature flags. But then you
accidentally enable a bunch of other flags!

(Also, we don't have a staging "environment" yet - it runs with
RAILS_ENV=production. So it is hard to differentiate from actual
production.)

Rather, this uses an environment variable, FEATURES_ENABLED, to enable
features at boot-time. On Heroku we will be able to set this properly
per-environment.