Add CRUD for CourtCaseSubscription model
This adds some creation/deletion of a CourtCaseSubscription model which
is how I'm thinking we'll represent a victim's relationship to court
cases. Although it is probably a 1:1 relation, we just had a call with
DA Underhill and he explicitly brought up the fact that it is actually
1:Many because a victim might be part of multiple case numbers if there
are multiple defendants (well, technically there can be many victims
too, so it's Many-to-Many I suppose).

No tests yet; under current assumptions this is going to be admin-only
functionality or being done automatically. I'll add tests as this
feature set solidifies.