Add first stab at the case show page
This is kind of a weird URL structure for now, but we'll resolve that
later.

For now, /subscription/:id goes to the show page for a single case. I
didn't get very close to the mocks for this because I'm on a flight
without internet. (Well technically they have internet, but charge
$39.95 to use it! That's sky-way robbery!)