Add ability to cache a search for DCJ offenders
I'm pretty sure this is going to create more problems than it solves,
but the use case is so that we can send users to a URL like:

  /offenders/dcj/123456 (where 123456 is the offender's SID)

But since we have to query the PPO lookup tool with both the SID and
last name, in order for us to re-fetch the information for that URL we
need to cache the last name.

There is an edge case around if you leave that tab open for too long
(longer than the cache result is valid), and refresh, then we need the
user to give us the last name again via search.