Add `jurisdiction` to OffenderScraper and OffenderGrouper
The `jurisdiction` will differentiate between SIDs from multiple
sources. I'm not entirely sure this is necessary because I believe that
all SIDs (from both a local *and* state-wide level) share a single
namespace... but this is the kind of thing that's really hard to change
later if the assumption turns out to be false.

Also it will allow us to differentiate content based on the jurisdiction
of the offender.