Add jurisdiction-specific offender show page
This creates a new show page for offenders, with an extra part of the
URL to namespace the offender by jurisdiction.

For example, previously an offender in state prison would have been at

  /offenders/1234

but now they are at

  /offenders/oregon/1234

and an offender under community supervision by Multnomah DCJ would be
at:

  /offenders/dcj/6789

To save a bit of pain converting all the URLs, I created a new helper
which defaults the jurisdiction to oregon. As this pattern solidifies,
I'll gradually remove it.

On the view side of things, there are a ton of conditionals for
rendering that page based on the jurisdiction of the offender. Not the
best pattern, but again, as this pattern and the design solidify, I will
plan to abstract better patterns out of this.