Convert `current_user` into variable
There was previously a helper `current_user` which used the
`@current_user` variable set in ApplicationController only if a
controller called the `current_user` controller method. This resulted in
transitioning between logged-in and logged-out depending on whether the
page called the `current_user` method in ApplicationController.

Instead, let's just always set the `@current_user` variable and use that
everywhere for consistency.