Upgrade oos-mechanizer gem
This gem extracts some common searching logic for OOS into a gem. I'm
using this in some other scraping scripts as well as in MyAdvocate.

This upgrade includes a more modular structure internal to the gem, as
well as forwarding through the error messages as exceptions. The error
with the `last_name` (fixed in the previous commit) would have been much
easier to detect with better error handling!