Ensure OffenderScraper supplements results with jurisdiction
This jurisdiction is important in grouping results and displaying the
link to the offender (/offenders/:jurisdiction/:sid). So, something
needs to include it. In the test suite I had stubbed data from
OffenderScraper that included the field, but never actually added the
field.

This adds the field and a regression spec.