Simply routes for offender related routes
This is a weird non-RESTful structure because we're using a "directory"
in the URL to filter the resources. Best I can tell, this is not a thing
that Rails encourages.

So this alters the routes to not even try to be RESTful. This will make
the route helper names easier.

Before:

                    GET  /offenders/:id(.:format)               redirect(301, /offenders/oregon/%{id})
 offender_offenders GET  /offenders/:jurisdiction/:id(.:format) offenders#show
   search_offenders POST /offenders/search(.:format)            offenders#search
          offenders GET  /offenders(.:format)                   offenders#index

After:
               GET      /offenders/:id(.:format)                redirect(301, /offenders/oregon/%{id})
 offender_path GET      /offenders/:jurisdiction/:id(.:format)  offenders#show
     offenders GET|POST /offenders(.:format)                    offenders#index

Regarding the removal of the `search_offenders` route, that route was
not being used anywhere. Instead, `offenders_path` was used.