Remove Notification model, controller, and views
This was originally going to track how we sent text messages inviting
users to MyAdvocate. However, it is now likely that we will re-introduce
this functionality in a totally different way. It's not a ton of code to
rewrite; it is probably easiest to just do that rather than try to
repurpose this code.