Update DcjSearch API to handle search by dob OR sid
The previous API required a two-step process to display an offender's
status:

1) Search by Last Name & SID
   -> This caches all search results, so then they can be referred to by
   SID only.
2) Navigate to /offenders/dcj/:sid
   -> This uses the search cache so as to avoid needing to encode the
   last name in the URL as well.

This is due to a restriction in how we are allowed to query the PPO
lookup tool.

We want to migrate the search UI to support searching by a DOB instead
of a SID in step 1. This became a bit complicated because I had
previously exploited the fact that SID was present in both steps to ease
caching. However, now, it is possible to find an offender without
knowing their SID.

So the two steps are still pretty much the same, but the DcjSearch API
has been rewritten so as to separate the steps into different methods.