Fix OffenderNameHighlighter to behave even if blank fields given
In the case where the user searches for a `last_name` but has a blank
(empty string) `first_name`, then the highlighter was attempting to
highlight all empty strings, which resulted in nothing being
highlighted. Thanks to the test coverage here, it's easy to verify that
adding a simple `presence` check does not break any other scenarios.