Move offender search to POST
Previously this was being done in the #show action using GET parameters.
This worked pretty well, except that it became difficult to determine
when the user had forgotten to fill out a form versus was just viewing
the form for the first time.

To help us be able to give actual form validation errors, this converts
the form to a POST and a separate controller action.

This also allows us to only search the intended jurisdiction.