Update DcjClient to accept a date object
This moves the responsibility for creating a Ruby date object from the
DcjClient into the OffenderJurisdictionsController. The controller is
now responsible for generating an error message if necessary, which
creates some complex control flow that could use another pass over to
simplify.

cc @tmaybe because we paired on this kind of