Add stretchy button for Offender Search buttons
These buttons need to be three columns wide (because if they are only 2
columns then the text wraps), which makes it impossible to center them
on a 12 column grid. So, I have taken the 3-column width and then
centered it with the good-'ol-margin-auto trick.

The buttons also need to know that they should expand to fill their
container, so I added a `stretch: true` option which makes them do that.