Change search button to "On Probation or Parole"
This updates the search button to read "On Probation or Parole", to
match the header. Although this has unfortunate side effects (that the
buttons all must be wider - 4 columns), I think it is better than
maintaining two buttons for the following reason:

I initially tried implementing this with two different pages - one for
probation and one for parole, and updating the headers to say "Look up
an offender on [probation/parole]" correctly.

However, there is no actual way for us to tell if an offender is being
supervised on probation, parole, or post-prison supervision, and it felt
weird in my local testing to be able to find probation offenders while
searching "for an offender on parole." Although, for a user, this is
unlikely to be a problem if they pick the right button, it still seems
strange to me to appear to offer two search pages that do different
things, when the underlying behavior is exactly the same.

Happy to hear other viewpoints on this though.

cc @pamdinevaCfA