Improve search page error appearance
This alters the search page errors to use a non-flash approach so the
site-wide flash approach can still be used for generic errors. Since the
errors are displayed between the page header and the page body, using a
controller-local variable for these errors will allow us to keep the
flash for general errors on other pages.