Give a 10 second quota for Dcj Baxter API to return
I was hoping that the production Baxter API would be faster, but it
looks like we still are dangerously close to the 15 second Heroku
timeout.

Although killing the thread in this way is likely to cause memory leaks,
it's way better than displaying an application error to the client.

TODO: set a Read timeout on the socket instead of using Timeout.timeout.