Add server-side event tracking when a user searches
Since the frontend is pretty light on interaction, this makes me believe
we'll be sending a lot of server-side events. This commit adds a
(possibly-overengineered) mixpanel framework on the server with the goal
of maintaining a consistent distinct_id for the user so the events can
be correlated.

On the first pageview, the user will have no mixpanel cookies and
mixpanel will assign a cookie for itself including an auto-generated
distinct id. (If we ever manually update the distinct id, this cookie
will be updated.)

This cookie will be passed to the server on subsequent pageviews in the
Cookie header. By parsing the header and saving its value in the
`request` object, we can have simpler controller code. By creating a
wrapper class, we can remove the responsibility for finding the viewer's
distinct_id from each `@mixpanel.track` call, since it is overwhelmingly
likely that the viewer is the user for which we wish to attribute the
event.

cc @tmaybe for feedback since you've done this separately