Add custom rspec matcher for mixpanel server-side events
Hey @tmaybe, this is a nifty pattern for allowing DRY specs that I have
seen work well before. By defining a custom matcher for mixpanel events,
most specs can declaratively specify their assertion of which events are
tracked -- so if the implementation of mixpanel changes later we can
just change it in the matcher.