Add styling for focused layout/feedback responses page
This adds a bunch of styling for these pages which gets them to look
pretty close to the mocks (except for some likely spacing discrepancies
and the implementation of the character counter on our feedback body.)

I tried to keep the concerns of the layout vs. the view itself
separated.