Exclude sandbox CSS from main application bundle
This uses a sprockets directive I hadn't heard about before to exclude
the sandbox CSS from the main application bundle, and include a new
top-level CSS bundle with only the sandbox styles on the sandbox page.
This has the effect that the sandbox page's styles will only apply on
the sandbox page itself.

Previously, the sandbox styles were included in the main application.css
bundle, which meant that they could override styles outside the sandbox.
This creates a problem when working with styles such as:

```css
.app-button {
  display: block;
}
```

which creates visual discrepancies across the site. This should not be
possible in a "sandbox" -- it should be safe to do whatever CSS you want
inside it.

cc @pamdinevaCfA -- heads up that the sandbox is now more of an actual
"sandbox"