Create `spinner` button and use it in Offender Search
This adds a new variant of the button, `spinner: true`, which adds a
`data-app-spinner` attribute to the button element. There is some light
javascript which intercepts clicks of those elements and then applies
the `app-button--loading` class, which does all the transition work. (I
also created a `loading: true` variant of the button so we could QA the
loaded state in the styleguide).

cc @pamdinevaCfA

Closes https://github.com/multnomah-fellowship/planning/issues/125