Create UtmAttribution model
This is the beginning of being able to track when a user completes an
action having entered the site with attribution parameters.

I'm not sure what this will look like in the frontend, but I'm thinking
of perhaps redirecting from ?utm_campaign=something to a URL without the
attribution parameters and storing the attribution ID in the session.