Add beta signup confirmation page
This adds a confirmation page with some happy messages if the user
decides to give us their email for the beta signup.

In doing so, this introduces a new pattern called a "pre-main focus"
area which is a text-centered content area that appears directly
underneath the site navigation. We will see if we can use it on some
other page.

Closes https://github.com/multnomah-fellowship/planning/issues/147