Add `disable_messages!` feature to Slack Client
Sometimes you just don't want the Slack client to send messages, like in
tests. Rather than making it opt-in for a test to try to send messages,
I decided an opt-out approach would make it more visible when a test
tried to send a notification. This means that the production behavior is
preserved unless the test author explicitly opts out, which I think will
make for better tests.