Add `RightsFlow` model for multi-page form
[The designs for the digital VRN form][1] call for it to be spread
across multiple pages with a sign up in the middle. Having encountered
this desire multiple times in my career, building stateful flows is
always a challenge. I've never seen a great approach, and I'm not sure
that this is one either.

This creates a `RightsFlow` model which is not persisted to the
database, rather it mixes in methods from `ActiveModel::Model`, which is
designed to just be enough methods to be usable in Rails view helpers.
Sweet. The `RightsFlow` model _could_ subclass some kind of generic
`Flow` concept, but that seemed like a premature abstraction to me if we
are only going to have a single flow.

I also wanted to keep the view logic as separated from the flow logic as
possible, since it is always the case that the visual design of
different stages of a flow will end up being just different enough to
cause headache.

This is still a WIP - it's an attempt at getting towards feature
completion before doing a gigantic styling pass.

cc @tmaybe

[1]: https://projects.invisionapp.com/share/GSCH646DC#/screens/242512758