Persist the Rights at the end of the flow
This fills in the implementation of the RightsFlow#persist! method, and
writes some tests to fit.

This represents a transition away from the `court_case_subscription`
model being tied to a user, since it is possible to fill out a VRN
without creating an actual account.

For now, I've made it actually create a `User` object with a blank
password, though it's likely we'll want to rename the
CourtCaseSubscription model to something else, like `VRNForm`.