Automatically log in the user when completing the rights flow
When the user completes the rights flow, an User account has been
created for them. (At least for right now. Later I think we will want to
change this.)

By setting the flow's `court_case_subscription_id` property after
creating the CourtCaseSubscription, the controller can log the user in
as the newly created user. This will help us handle the condition of
pressing the back button from the confirmation page.