Separate user from CourtCaseSubscription
If the person subscribing to a court case does not want to create an
account, we should store their personal identifiable information on the
CourtCaseSubscription object. Later, we can create a User object if they
want to sign in.

On CourtCaseSubscription, that means that we need some custom validation
to make sure that either there is a user _or_ there is contact
information via email or phone. (Otherwise, how have you subscribed?!)

This also lumps in some visual changes to the confirmation page
depending on whether the subscription was "just created" or not, as
given by the `created_at` timestamp equaling the `updated_at` timestamp.