Fix all other Rubocop lints
This moves pretty much all the other cops (the ones that require manual
intervention) into the .rubocop.yml. I disabled a bunch of the Metrics
ones, since although they point to smelly code, I don't think it is the
linter's place to identify it. (It should be in code review.)

This also fixes a couple tests that were broken by the refactors.