Test for actual data in OffenderSearchCache
In the case of a race condition which results in the call to
`OffenderSearchCache.update_attributes` failing, an OffenderSearchCache
record is persisted with a blank data field.

Although the race would be good to be eliminate, it suffices to check
for the data to gracefully recover from the problem at first.