Standardize Right names to "flag_x" and give them keys
I'd like to be able to refer to the name of the right ("A-DDA to
assert[...]") with a symbol like Right::RIGHTS[:flag_a] rather than
Right::NAMES[0]. This will also allow us to remove the RIGHTS_MAPPING
constant if all the field names line up. (Left for a subsequent commit.)

This also makes sure that all the rights are represented in the rights
flow.