Add first version of the VRN email
(VRN = Victim Rights Notification)

As part of implementing the VRN flow digitization[1], we need to
actually create real-world evidence that the user has gone through the
flow. This takes two parts, both having to do with email.

First, we need to email the victim to confirm the rights that they
selected. (This commit.)

Next, we need to email their advocate so that the advocate can update
the record in their backend database system.

This commit uses a couple nice Rails ecosystem features -- ActionMailer
previews and premailer to automatically inline CSS. This allows us to
share our styling between the rest of the site and this email.

It's likely that this will take many iterations to look good in mail
clients. Here goes nothin...

[1]: https://github.com/multnomah-fellowship/planning/issues/162