Set ActionMailer asset_host and raise_delivery_errors in prod
`asset_host` is necessary so that the `image_path` helper will output an
absolute URL instead of a relative one.

`raise_delivery_errors` could result in poor user experience, but I'd
say it's crucial for us to know if these messages are not being sent
properly during development. Later, we can turn this off if it gets
noisy.