Add RightsMailer.vrn_advocate_update email
This is an email that goes to the DA Advocate when the user updates or
initially sets their rights. This will need another design pass; I just
want to get this in for now to unblock our usability tests next week.

Also something to figure out is where these emails will go - I'd like to
not receive them forever.