Use the first entry in X-Forwarded-For header for mixpanel
The mixpanel server-side events are not from the same location as the
client side events, probably due to cloudflare being in the middle.

To do this, we'll use whatever IP address the client gives first in
X-Forwarded-For. I gave it a scary name because it can be spoofed by
clients.