Support both new and old product names
For now, until we have the name change redirect in place, it will be
helpful to support both the old and new names for the product depending
on the configured domain name for the application install.

To keep this logic in one place under control I created a helper with
good test coverage.