Add offender search pages to our sitemap
This will help Google find the search pages. I tweaked their <title>
tags a little bit so that they are a bit more googlable based on a vague
reading of the results from our Adwords campaign, which showed good
results for the words "inmate" and "find" instead of "offender" and
"look up".