Add starting point of baking an AMI
The goal of this line of work is to create a repeatable build process
that is easy to change without fear of breaking things. Multnomah County
IT is willing to spin up any AMI that we hand them, so let's create one
that has:

1. All the dependencies installed
2. A specific version of the code
3. A specific version of configuration

This is the first time I have tried to make an image with both code and
configuration baked in. Here, it feels right because it is a single
artifact in totality that IT just has to hit "Run" on and then it will
work.

As for the data storage, this is another concern -- I need to
investigate the possibility of using RDS.