Add "become: true" to most tasks
Almost all tasks are run from a superuser privileges assumption; this
commit adds "become: true" to prefix everything with sudo.

The one gotcha is that the first "install python" task needs to be
conditionally sudo'd depending on if sudo exists. If sudo doesn't exist,
then we have to assume that the command will run as a root user anyway.
(Since we cannot install "sudo" until python is installed.)