Add ruby deployment modules from Ansible Galaxy
Some dude wrote a trio of modules that are useful in deploying Rails.
This will need some additional thought to handle the two cases of 1)
building an AMI that contains a copy of the code and configuration and
2) also allowing a deployer to log into said VM and run the deployment
steps to start a new version of code.

Although the second case is an antipattern in an immutable AWS world,
it's unfortunately probably the world we live in whilst Multnomah County
IT is holding the keys to the AWS kingdom. (Since we would either need
to involve a human or create a permission sharing scheme to handle the
provisioning of new VMs if we only deployed by rolling out new AMIs.)

This commit doesn't quite work yet -- the `bundle install` step fails
with an error "[Errno 2] No such file or directory" right after
ansible announces the warning "[WARNING]: could not parse environment
value, skipping: [u'command_environment']", which seems suspicious and
likely to cause the problem if the PATH were not set to include the
`bundle` executable.