Get ansible to a passable state in Docker
There's still some work to do to make the ansible provisioning work
properly on Amazon, but this gets very very close.

The trick was to upgrade the `nicolai86.rails-deployment` module to the
latest version on `master` that has not yet been released to the Ansible
Galaxy.

After building, this container can be started (with Docker for Mac) with
the following command:

docker run -ti -p 3000:3000 --cap-add=SYS_ADMIN -e container=docker \
  -v /sys/fs/cgroup:/sys/fs/cgroup \
  multnomah-fellowship/casecompanion:latest /sbin/init

...which will start nginx and the Rails web process.

Still some TODOs but this gives a good starting point!