Add rbenv cache for Docker builds
Since it takes ~10 min to build Ruby when building the docker image, to
mitigate this, we can bind-mount the rbenv directory into docker.

However, since this is unexpected behavior, I don't want to enable it
by default, instead opting in via the `cache_rbenv_during_build` extra
var in `variables.json`.

Note that there is no way to conditionally add a mount to Docker in
packer, so the folder is mounted regardless and it's just not used
without setting cache_rbenv_during_build=true.

When running the container with systemd, you also have to bind-mount in
rbenv, making the combined systemd command:

docker run -ti -p 3000:3000 --rm \
  --cap-add=SYS_ADMIN -e container=docker \
  -v /sys/fs/cgroup:/sys/fs/cgroup \
  -v ~/.cache/docker-casecompanion/rbenv:/home/app/.rbenv \
  multnomah-fellowship/casecompanion:latest /sbin/init