Add self-signed nginx TLS configuration
This is important so that whatever load balancer / CDN lies in front of
nginx can communicate with nginx "securely". More secure would be an
actually valid TLS certificate, but this will do the trick for now until
we are big enough to roll our own PKI.

This also helps us keep Rails's `config.force_ssl` enabled in
production, a win for security through HSTS and HTTP-Only cookies.