Add capistrano for deployment
This is the defacto way to deploy a Rails app, and is essentially what
the Ansible module `rails-deployment` re-implements. I would prefer to
re-use the Ansible module for the deployment, except that we are forced
into the deploying machine being a Windows machine, and Ansible does not
run on Windows.

But Capistrano does!