Update CRIMES ETL scripts for new Windows box
One-off scripts never work on the second host you try them on, and these
are no exception -- when we got access to a new Windows VM it revealed a
bunch of assumptions in this script.

First, primarily, the 'sqlcmd' executable must be run as the
'mcda\doonerto' user which is not the user I am logged in to on the
Multnomah county network. Hence the creation of 'dump-data-wrapper.bat'
which sets the permissions correctly for the inner script (dump-data.bat).
I'm not sure how to automate this.

Secondly, I had forgotten to commit the 'import-data.sh' script and also
the 'q-before.sql' query file.

Thirdly, permissions look a bit different in the actual production deploy
environment, so I created a '.psqlenv' file that has the
PGUSER/PGPASSWORD/etc. environment variables defined for purposes of these
scripted commands.

Fourthly, and finally, I removed the dependency on the Python 'csvkit'
library, which although convenient, seemed a bit too onerous to install on
application servers just to import SQL from a CSV file. This means that we
have to check-in the SQL which creates these intermediate tables... but
that is probably for the better so it is more explicit.

cc @lkogler