Add in-app Email link tracking
This adds the `ahoy_email` gem which is a simple gem that adds tracking
parameters to links in outgoing emails and records the results in a
database table. It supports arbitrary subscriptions as well, so I've
wired it up to mixpanel so we can capture the case where the same user
opens or clicks the same email multiple times. (The database table just
has a single `opened_at` timestamp.)