Add RightsPdfGenerator to create the PDF
This module takes as input a CourtCaseSubscription and outputs a PDF's
contents. (Rather than generating a file on disk, I believe outputting
the contents themselves will facilitate attaching to an email.)

The prawn DSL is sadly quite verbose for a simple layout that we need,
but at least we don't need to install pdftk and get a fillable version
of the form from the DA to do this.

Also included, for ease of development, is an endpoint which renders the
PDF into a browser response. It's at:

  http://localhost:3000/rights/preview.pdf