Add a <select> full of advocates on Rights confirmation page
In order for us to send email to a victim's advocate, we need to know
who they are. This adds a "Advocate Name" dropdown on the final page of
the rights flow, which persists the Email of the advocate in the
CourtCaseSubscription object.

The data for this is checked-in as a CSV, although eventually this
should be either managed in-app or automatically fetched at deploy-time.