Fix accidentally filtering out Phone-less users from query
We are intending to filter the set of users we are sending VRNs to by
three criteria:

1) the case was created in the past two months
2) a VRN was filled out by the users
3) the user is not a test user

(these are all filters in 01_VRN.sql).

Then, we further filter down the results to only users with email
addresses.

However, because an earlier version of the email contained the desire to
remind the users that they could update their phone number, I had also
pulled over the `WHERE` clause from Karl's query which filters out phone
numbers that are not "default" (whatever that means).

For one specific user with only a non-"default" phone number, because of
how SQL JOINs work the OUTER JOIN was succeeding and returning the row,
but then the row was filtered out by the WHERE. Instead, we want the
OUTER JOIN to just fail and return NULL values for those columns.