Add "confirmation" step to VRN flow
This adds a new step in the VRN flow (the `RightsFlow` object) which
requires the user to "sign" a confirmation that they picked the correct
rights.

The signature is simply their name and a check box affirming that the
rights look correct. (We received sign-off that this is legally sound.)