Validate that the case number looks good
Some case numbers start with "Y" for youth (we don't support those yet).
And some are suffixed with -DV, -A, -C, etc., depending on the unit
handling the case (those suffixes should be omitted).

Also, in case someone puts in the Court case number (e.g. 17CR1234),
this gives a good error as well.