Remove concept of skipping a RightFlow step
Initially I had thought we would want to skip the `create_account` step
so that you don't see it when you hit the browser "Back" button. But...
what if you mistype your email or name?! It makes no sense to not be
able to edit it after-the-fact.

Removing the `skip_step?` method also simplifies some other logic, which
I'm glad is tested.