Reset VRN flow after hitting the "done" page
This is pretty tricky -- we want to make it so once you see the "done"
step in the VRN flow, you have made it through the VRN flow and clicking
"back" clears your browser state so that you start the flow again. This
also implies the ability for the flow to redirect you to the proper
place, if you are viewing it out-of-order.

To do this, I added a `redirect_to_page` flow method that the controller
uses to determine whether to redirect the user based on a new
`completed_step` variable that is tracked by the flow itself with the
value of the latest step that was successfully completed (i.e. not
invalid).

This complicates the controller tests, so I removed them since the
request specs cover it as well.

Also, I tried to make the logic around clearing the session a bit more
clear. Either the cookie is updated or cleared based on the value of
`should_delete_flow?`.